+++ lib/doggo/storybook/cluster.ex
@@ -23,6 +23,15 @@ defmodule Doggo.Storybook.Cluster do
       <div>One</div>
       <div>Two</div>
       <div>Three</div>
+      <div>Four</div>
+      <div>Five</div>
+      <div>Six</div>
+      <div>Seven</div>
+      <div>Eight</div>
+      <div>Nine</div>
+      <div>Ten</div>
+      <div>Eleven</div>
+      <div>Twelve</div>
       """
     ]
   end